requirements.txt
anyio==4.0.0
black==24.4.2
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
coloredlogs==15.0.1
coverage==7.5.1
cryptography==41.0.3
distlib==0.3.7
editables==0.5
filelock==3.12.4
h11==0.14.0
hatch==1.7.0
hatchling==1.18.0
httpcore==0.18.0
httpx==0.25.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pexpect==4.8.0
platformdirs==3.10.0
pluggy==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pygame-ce==2.4.1
Pygments==2.16.1
pyperclip==1.8.2
pytest==8.1.2
pytest-cov==5.0.0
pytest-mock==3.14.0
rfc3986==2.0.0
rich==13.5.2
SecretStorage==3.3.3
shellingham==1.5.1
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli_w==1.0.0
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.11.0
userpath==1.9.1
virtualenv==20.24.5
zipp==3.16.2
